Render Homepage elements in App
