+++ src/App.js
@@ -1,13 +1,13 @@
 import React, { useState, useEffect, Fragment } from 'react'
-import { Route, Routes } from 'react-router-dom';
-import Homepage from './Components/Homepage/Homepage'
-import Header from './Components/Header/Header'
+import { Link, Route, Routes } from 'react-router-dom';
+import Homepage from './Components/Homepage/Homepage';
+import Header from './Components/Header/Header';
 import './App.scss';
 import MatchedBeers from './Components/MatchedBeers/MatchedBeers';
 
 const App = () => {
   const [data, setData] = useState([]);
-  const [error, setError] = useState([])
+  const [error, setError] = useState([]);
 
   const fetchAllBeers = async () => {
     try {
@@ -24,15 +24,30 @@ const App = () => {
     fetchAllBeers()
   }, [])
 
-
+  const filteredTypes = (type) => {
+    const beerList = data.filter((beer) => {
+      if(beer.name.toLowerCase().includes(type) || beer.tagline.toLowerCase().includes(type)) {
+        return beer
+      }
+    })
+    return beerList
+  }
 
   return (
     <div className="App">
-      <Header />
-      <Routes>
-        <Route path='/' element={<Homepage beerData={data}/>} />
-        <Route path='/beers/type' element={<MatchedBeers/>}/>
-      </Routes>
+    <Header />
+    <section className="homepage">
+      <div className="filter-container">
+        <p className="start-text">let's get started</p>
+        <div className="filter-links">
+        <p>find beers by the food they pair with
+        <Link to={`/beers`}/>
+        </p>
+        <article>
+        </article>
+        </div>
+      </div>
+    </section>
     </div>
   )
 }
