+++ src/Components/Homepage/Homepage.css
@@ -1,32 +1,1 @@
-.homepage {
-  border: 2px solid blue;
-  height: 85%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-}
-
-.filter-container {
-  border: 3px solid purple;
-  width: 75vw;
-  height: 85%;
-}
-
-.start-text {
-  font-size: 2em;
-  text-align: center;
-}
-
-.filter-links {
-  border: 2px solid white;
-  height: 90%;
-  width: 90%;
-  margin: 0 auto;
-}
-/*# sourceMappingURL=Homepage.css.map */
+/* No CSS *//*# sourceMappingURL=Homepage.css.map */
