+++ src/Components/Homepage/Homepage.js
@@ -1,30 +1,27 @@
-import React, {useEffect,useState } from 'react';
-import {Link, NavLink} from 'react-router-dom'
-import './Homepage.scss'
+// import React, {useEffect,useState } from 'react';
+// import MatchedBeers from '../MatchedBeers/MatchedBeers'
+// import {Link, NavLink} from 'react-router-dom'
+// import './Homepage.scss'
 
 
-const Homepage = ( {beerData} ) => {
-  const filteredTypes = (type) => {
-    const beerList = beerData.filter((beer) => {
-      if(beer.name.toLowerCase().includes(type) || beer.tagline.toLowerCase().includes(type)) {
-        return beer
-      }
-    })
-    return beerList
-  }
+// const Homepage = () => {
 
-  return (
-    <section className="homepage">
-      <div className="filter-container">
-        <p className="start-text">let's get started</p>
-        <div className="filter-links">
-        <p>find beers by the food they pair with</p>
-        </div>
-      </div>
-    </section>
-  )
+//   return (
+//     <section className="homepage">
+//       <div className="filter-container">
+//         <p className="start-text">let's get started</p>
+//         <div className="filter-links">
+//         <p>find beers by the food they pair with
+//         <Link to={`/beers`}/>
+//         </p>
+//         <article>
+//         </article>
+//         </div>
+//       </div>
+//     </section>
+//   )
 
-}
+// }
 
 
 
@@ -42,4 +39,4 @@ const Homepage = ( {beerData} ) => {
 
 
 
-export default Homepage
+// export default Homepage
