+++ src/Components/Homepage/Homepage.scss
@@ -1,25 +1,25 @@
-.homepage{
-  border: 2px solid blue;
-  height: 85%;
-  display:flex; 
-  justify-content: center;
-  align-items:center;
-}
+// .homepage{
+//   border: 2px solid blue;
+//   height: 85%;
+//   display:flex; 
+//   justify-content: center;
+//   align-items:center;
+// }
 
-.filter-container{
-  border: 3px solid purple;
-  width: 75vw;
-  height: 85%;
-}
+// .filter-container{
+//   border: 3px solid purple;
+//   width: 75vw;
+//   height: 85%;
+// }
 
-.start-text{
-  font-size: 2em;
-  text-align: center;
-}
+// .start-text{
+//   font-size: 2em;
+//   text-align: center;
+// }
 
-.filter-links{
-  border: 2px solid white;
-  height: 90%;
-  width: 90%;
-  margin: 0 auto;
-}
+// .filter-links{
+//   border: 2px solid white;
+//   height: 90%;
+//   width: 90%;
+//   margin: 0 auto;
+// }
