+++ src/Components/MatchedBeers/MatchedBeers.js
@@ -1,7 +1,8 @@
 import React, {useEffect,useState } from 'react';
 import './MatchedBeers.scss'
 
-const MatchedBeers = () => {
+const MatchedBeers = ({ filteredBeers }) => {
+  console.log(filteredBeers)
   return (
     <h1>MatchedBeers!</h1>
   )
